+++ tests/test_context/test_requires_context_io_result/test_requires_context_io_result_cast.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from returns.context import RequiresContextIOResult, RequiresContextIOResultE
+from returns.context import (
+    ReaderIOResult,
+    ReaderIOResultE,
+    RequiresContextIOResult,
+    RequiresContextIOResultE,
+)
 
 
 def _function(arg: int) -> RequiresContextIOResultE[int, float]:
@@ -12,6 +17,19 @@ def _function(arg: int) -> RequiresContextIOResultE[int, float]:
 
 
 def test_requires_context_io_resulte():
-    """Ensures that RequiresContextResultE correctly typecast."""
+    """Ensures that RequiresContextIOResultE correctly typecast."""
     container: RequiresContextIOResult[int, float, Exception] = _function(1)
+    assert container(0) == RequiresContextIOResult.from_success(10.0)(0)
+
+
+def test_requires_context_io_aliases():
+    """Ensures that ReaderIOResult correctly typecast."""
+    container: ReaderIOResultE[int, float] = _function(1)
+    container2: ReaderIOResult[int, float, Exception] = _function(1)
+    container3: ReaderIOResultE[int, float] = ReaderIOResultE.from_success(
+        10.0,
+    )
+    container4: ReaderIOResultE[int, float] = ReaderIOResult.from_success(10.0)
+
+    assert container(0) == container2(0) == container3(0) == container4(0)
     assert container(0) == RequiresContextIOResult.from_success(10.0)(0)